models/ requirement.txt
tensorflow-gpu>=2.2.0
